[CL-2101] Moved stat_user_policy patch to core
