+++ back/app/policies/stat_user_policy.rb
@@ -18,7 +18,15 @@ class StatUserPolicy < ApplicationPolicy
     private
 
     def resolve_for_active
-      user.admin? ? scope.all : scope.none
+      user.admin? ? scope.all : resolve_for_project_moderator
+    end
+
+    def resolve_for_project_moderator
+      return scope.none unless user.project_moderator?
+
+      ::UserRoleService.new.moderatable_projects(user)
+        .map { |project| ::ProjectPolicy::InverseScope.new(project, scope).resolve }
+        .reduce(:or)
     end
   end
 
@@ -67,8 +75,6 @@ class StatUserPolicy < ApplicationPolicy
   end
 
   def show_stats_to_active?
-    admin?
+    admin? || user.project_moderator?
   end
 end
-
-StatUserPolicy.prepend_if_ee('ProjectManagement::Patches::StatUserPolicy')
