--- back/engines/commercial/project_management/app/policies/project_management/patches/stat_user_policy.rb
@@ -1,29 +1,0 @@
-# frozen_string_literal: true
-
-module ProjectManagement
-  module Patches
-    module StatUserPolicy
-      def self.prepended(base)
-        base::Scope.prepend(Scope)
-      end
-
-      module Scope
-        def resolve_for_active
-          super.or(resolve_for_project_moderator)
-        end
-
-        def resolve_for_project_moderator
-          return scope.none unless user.project_moderator?
-
-          ::UserRoleService.new.moderatable_projects(user)
-            .map { |project| ::ProjectPolicy::InverseScope.new(project, scope).resolve }
-            .reduce(:or)
-        end
-      end
-
-      def show_stats_to_active?
-        admin? || user.project_moderator?
-      end
-    end
-  end
-end
